+++ lib/assert.js
@@ -52,6 +52,8 @@ const ERR_DIFF_DEACTIVATED = 0;
 const ERR_DIFF_NOT_EQUAL = 1;
 const ERR_DIFF_EQUAL = 2;
 
+let warned = false;
+
 // The assert module provides functions that throw
 // AssertionError's when particular conditions are not met. The
 // assert module must conform to the following interface.
@@ -80,8 +82,18 @@ function fail(actual, expected, message, operator, stackStartFn) {
   } else if (argsLen === 1) {
     message = actual;
     actual = undefined;
-  } else if (argsLen === 2) {
-    operator = '!=';
+  } else {
+    if (warned === false) {
+      warned = true;
+      process.emitWarning(
+        'assert.fail() with more than one argument is deprecated. ' +
+          'Please use assert.strictEqual() instead or only pass a message.',
+        'DeprecationWarning',
+        'DEP00XXX'
+      );
+    }
+    if (argsLen === 2)
+      operator = '!=';
   }
 
   innerFail({
